+++ src/js/library/clear-containers.js
@@ -1,0 +1,7 @@
+import { refs } from './library-refs';
+
+const clearLibraryContainer = () => {
+  refs.library.innerHTML = '';
+};
+
+export { clearLibraryContainer };
