+++ src/js/library/library-events.js
@@ -1,0 +1,5 @@
+import { activateWatchedBtn, activateQueueBtn, deactivateWatchedBtn, deactivateQueueBtn } from './library-buttons-functions';
+import { refs } from './library-refs';
+import { clearLibraryContainer } from './clear-containers';
+
+//колбеки для кнопок
